Pull alias parsing into helper
