+++ extensions/terminal-suggest/src/shell/zsh.ts
@@ -3,9 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as vscode from 'vscode';
 import type { ICompletionResource } from '../types';
-import { execHelper, spawnHelper } from './common';
+import { execHelper, getAliasesHelper } from './common';
 import { type ExecOptionsWithStringEncoding } from 'node:child_process';
 
 export async function getZshGlobals(options: ExecOptionsWithStringEncoding, existingCommands?: Set<string>): Promise<(string | ICompletionResource)[]> {
@@ -22,29 +21,5 @@ async function getBuiltins(options: ExecOptionsWithStringEncoding, existingComma
 }
 
 async function getAliases(options: ExecOptionsWithStringEncoding): Promise<ICompletionResource[]> {
-	// This must be run with interactive, otherwise there's a good chance aliases won't
-	// be set up. Note that this could differ from the actual aliases as it's a new bash
-	// session, for the same reason this would not include aliases that are created
-	// by simply running `alias ...` in the terminal.
-	const aliasOutput = await spawnHelper('zsh', ['-ic', 'alias'], options);
-	const result: ICompletionResource[] = [];
-	for (const line of aliasOutput.split('\n')) {
-		const match = line.match(/^(?<alias>[a-zA-Z0-9\.:-]+)=(?:'(?<resolved>.+)'|(?<resolved>.+))$/);
-		if (!match?.groups) {
-			continue;
-		}
-		let definitionCommand = '';
-		let definitionIndex = match.groups.resolved.indexOf(' ');
-		if (definitionIndex === -1) {
-			definitionIndex = match.groups.resolved.length;
-		}
-		definitionCommand = match.groups.resolved.substring(0, definitionIndex);
-		result.push({
-			label: match.groups.alias,
-			detail: match.groups.resolved,
-			kind: vscode.TerminalCompletionItemKind.Alias,
-			definitionCommand,
-		});
-	}
-	return result;
+	return getAliasesHelper('zsh', ['-ic', 'alias'], /^(?<alias>[a-zA-Z0-9\.:-]+)=(?:'(?<resolved>.+)'|(?<resolved>.+))$/, options);
 }
